[Move `@kbn/config-schema` to server] sharedux plugins
