+++ src/plugins/no_data_page/public/plugin.ts
@@ -8,7 +8,7 @@
 
 import type { CoreSetup, CoreStart, Plugin, PluginInitializerContext } from '@kbn/core/public';
 import type { NoDataPagePublicSetup, NoDataPagePublicStart } from './types';
-import type { NoDataPageConfig } from '../config';
+import type { NoDataPageConfig } from '../server/config';
 
 export class NoDataPagePlugin implements Plugin<NoDataPagePublicSetup, NoDataPagePublicStart> {
   constructor(private initializerContext: PluginInitializerContext<NoDataPageConfig>) {}
