+++ src/plugins/no_data_page/server/config.ts
