+++ src/plugins/share/server/config.ts
